Always set REMOTE_USER when valid auth provided (even if not required).
Interpret "auth not required" (i.e. auth_required_callback returning False)
to mean: missing/invalid auth will not result in a 40x error,
but if a client provides valid auth, it will now be exposed to the app.

This better supports cases where auth is not required for a particular request
(e.g. a health check endpoint hit anonymously by Kubernetes),
but it is still useful to know who made the request when valid auth is provided
(e.g. a script or user interactively hitting the health check endpoint
from a client that does pre-emptive Kerberos).